chore: Move route paths from coordinator to routes file
